adapters/redis: Rename to wredis
