+++ adapters/wredis/store_test.go
@@ -1,4 +1,4 @@
-package redis
+package wredis
 
 import (
 	"testing"
